finish module create item

+++ pkg/itemManaging/model/itemManaging.go
@@ -6,7 +6,7 @@ type (
 		Name        string `json:"name" validate:"required,max=64"`
 		Description string `json:"description" validate:"required,max=128"`
 		Picture     string `json:"picture" validate:"required"`
-		Price       uint64 `json:"price" validate:"required,max=64"`
+		Price       uint   `json:"price" validate:"required,max=64"`
 	}
 
 	ItemEditingReq struct {
@@ -14,6 +14,6 @@ type (
 		Name        string `json:"name" validate:"omitempty,max=64"`
 		Description string `json:"description" validate:"omitempty,max=128"`
 		Picture     string `json:"picture" validate:"omitempty"`
-		Price       uint64 `json:"price" validate:"omitempty,max=64"`
+		Price       uint   `json:"price" validate:"omitempty,max=64"`
 	}
 )
